apps/backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.2.0
celery==5.4.0
redis==5.0.8
Pillow==10.4.0
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn>=21.2.0
celery>=5.3,<6
django-celery-beat>=2.6,<3
django-celery-beat>=2.6,<3
whitenoise==6.7.0
